Final changes and uploads for evaluations
evaluation_on_gan is the main code to generate the evaluation metrics on the paper, FCN_facades was used to train an FCN for facades image segmentation using transfer learning, facades dataset is used to create a dataloader for that FCN, and changes were made to evaluate_dataset to sort listed directories for easier lookup of generated images. The file for loading the 200 epoch fcn weights was too large for github so it is omitted. In addition, all the code that failed to use pre-trained FCN is omitted